+++ tests/Models/AutoPost.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Hcs\LaraCache\Tests\Models;
+
+/** A Post that caches every read, regardless of the global mode. */
+class AutoPost extends Post
+{
+    protected $table = 'posts';
+
+    protected $cacheMode = 'auto';
+}